Add runnable script for creating Hubs from curated list
